+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * StatusCode to errors
 * Check for register to see if user already exists
 * Easier way import custom errors
+* Support for passing entire header into validate
 
 ### Changed
 * Have InvalidRoleError extend ForbiddenError to return same information
 * Login method returns the user
+* Change time expiration to number in seconds vs string
 
 ## [1.0.0] - 2020-12-14
 ### Added